refactor: split tool-use core into srp modules
